math_quiz/requirements.txt
Django==3.2.7
gunicorn==20.0.4
django-environ==0.4.5
psycopg2-binary==2.8.5
djangorestframework==3.12.4
markdown==3.3.4
django-filter==21.1
Pillow==8.4.0
python-dotenv==0.19.2